use manually scoped non-scoped enum.
no typecasts are required and still not polluting the namespace -> okay
solution.

+++ hardware/src/barometer/ms5837.cpp
@@ -50,13 +50,12 @@ bool MS5837::Reset() {
 
 MS5837::Status MS5837::Read(MS5837::Oversampling oversampling) {
   using namespace std::chrono;
-  double tmp = 2.2e-6 * (0x01 << (8 + static_cast<int>(oversampling)));
+  double tmp = 2.2e-6 * (0x01 << (8 + oversampling));
   auto conversion_time = round<nanoseconds>(duration<double>(tmp));
 
   std::array<uint8_t, 0> data;
-  i2c::WriteData(
-      file_handle_, kAddress,
-      kRegisterConvertTemperature + 2 * static_cast<int>(oversampling), data);
+  i2c::WriteData(file_handle_, kAddress,
+                 kRegisterConvertTemperature + 2 * oversampling, data);
   std::this_thread::sleep_for(conversion_time);
   if (!i2c::ReadData(file_handle_, kAddress, kRegisterAdcRead,
                      temperature_raw_)) {
@@ -64,8 +63,7 @@ MS5837::Status MS5837::Read(MS5837::Oversampling oversampling) {
   }
 
   i2c::WriteData(file_handle_, kAddress,
-                 kRegisterConvertPressure + 2 * static_cast<int>(oversampling),
-                 data);
+                 kRegisterConvertPressure + 2 * oversampling, data);
   std::this_thread::sleep_for(conversion_time);
   if (!i2c::ReadData(file_handle_, kAddress, kRegisterAdcRead, pressure_raw_)) {
     return Status::kIOError;
